working tests against preview deployment when run locally

+++ e2e/global.teardown.ts
@@ -1,22 +1,25 @@
+/* eslint-disable no-process-env */
 import { test as teardown } from '@playwright/test';
 import { execSync } from 'child_process';
 import { UTApi } from 'uploadthing/server';
 
 teardown('delete test database', async () => {
-  // remove uploaded files from uploadthing
-  // eslint-disable-next-line no-console
-  console.log('🗑️ Deleting uploaded files from uploadthing');
-  const utapi = new UTApi({
-    // TODO: figure out why we cannot use getUTApi here
-    // eslint-disable-next-line no-process-env
-    token: process.env.E2E_UPLOADTHING_TOKEN,
-  });
+  if (!process.env.CI) {
+    // remove uploaded files from uploadthing
+    // eslint-disable-next-line no-console
+    console.log('🗑️ Deleting uploaded files from uploadthing');
 
-  await utapi.listFiles({}).then(({ files }) => {
-    const keys = files.map((file) => file.key);
+    const utapi = new UTApi({
+      // TODO: figure out why we cannot use getUTApi here
+      token: process.env.UPLOADTHING_TOKEN,
+    });
+
+    await utapi.listFiles({}).then(({ files }) => {
+      const keys = files.map((file) => file.key);
       return utapi.deleteFiles(keys);
-  });
+    });
 
-  // Stop and remove test db
-  execSync('docker compose -f docker-compose.test.yml down -v', { stdio: 'inherit' });
+    // Stop and remove test db
+    execSync('docker compose -f docker-compose.test.yml down -v', { stdio: 'inherit' });
+  }
 });
